+++ src/hin/master.h
@@ -6,13 +6,24 @@
 #include <sys/socket.h>
 #include <sys/uio.h>
 
-enum { HIN_DAEMONIZE = 0x1 };
+enum {
+HIN_QUIT = 0x1, HIN_DAEMONIZE = 0x2, HIN_PRETEND = 0x4, HIN_RESTARTING = 0x8,
+};
+
+typedef struct hin_master_sock_struct {
+  int sockfd;
+  int type;	// listen/connection, udp/tcp, etc
+  intptr_t ai_family, ai_protocol, ai_socktype;
+  struct sockaddr ai_addr;
+  socklen_t ai_addrlen;
+  void * server;
+  struct hin_master_sock_struct * next;
+} hin_master_socket1_t;
 
 typedef struct {
   int sockfd;
-  int type;	// listen/connection, udp/tcp, etc
-  struct sockaddr in_addr;
-  socklen_t in_len;
+  struct sockaddr ai_addr;
+  socklen_t ai_addrlen;
 } hin_master_socket_t;
 
 typedef struct {
@@ -24,22 +35,22 @@ typedef struct {
 
 typedef struct {
   int id;
+  int quit;
+  uint32_t flags;
   uint32_t debug;
-  char * exe_path;
-  char * conf_path;
+  const char * exe_path;
+  const char * conf_path;
   const char * logdir_path;
   const char * cwd_path;
   const char * pid_path;
-  uint32_t flags;
-  int sharefd;
-  int quit;
   int restart_pid;
-  int restarting;
   void * servers;
+  hin_master_socket1_t * socket, * last_socket;
   int num_client;
   int num_connection;
   hin_client_t * server_list;
   hin_client_t * connection_list;
+  int sharefd;
   hin_master_share_t * share;
 } hin_master_t;