+++ src/http/server/httpd_file.c
@@ -261,7 +261,7 @@ int httpd_handle_file_request (hin_client_t * client, const char * path, off_t p
   buf->ssl = &client->ssl;
   buf->debug = http->debug;
 
-  hin_server_blueprint_t * socket = http->c.parent;
+  hin_server_t * socket = http->c.parent;
   hin_server_data_t * data = socket->c.parent;
 
   if (HIN_HTTPD_ASYNC_OPEN) {